+++ sdkApiVideo/API/RequestsBuilder.swift
@@ -11,7 +11,7 @@ public class RequestsBuilder{
     private func genericUrLRequestBuilder(apiPath: String, tokenType: String, key: String, httpMethod: String) -> URLRequest{
         var request = URLRequest(url: URL(string: apiPath)!)
         request.setValue("\(tokenType) \(key)", forHTTPHeaderField: "Authorization")
-        request.setValue("api.video SDK (ios; v:0.1.5; )", forHTTPHeaderField: "User-Agent")
+        request.setValue("api.video SDK (ios; v:0.1.6; )", forHTTPHeaderField: "User-Agent")
         request.httpMethod = httpMethod
         return request
     }
@@ -54,7 +54,7 @@ public class RequestsBuilder{
     }
     public func urlSessionBuilder() -> URLSession {
         let sessionConfig = URLSessionConfiguration.default
-        sessionConfig.httpAdditionalHeaders = ["User-Agent": "api.video SDK (ios; v:0.1.5; )"]
+        sessionConfig.httpAdditionalHeaders = ["User-Agent": "api.video SDK (ios; v:0.1.6; )"]
         let session = URLSession(configuration: sessionConfig)
         return session
     }