+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All changes to this project will be documented in this file.
 
 ## [0.1.6] - 2021-06-15
 - Change return type on upload methodes
+- Add LiveStream 'public' attribute
 
 ## [0.1.5] - 2021-02-09
 - Change upload large file method to stream